requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
configparser==5.0.2
Flask==1.1.2
gunicorn==20.1.0
hashids==1.3.1
idna==2.10
importlib-metadata==4.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
psycopg2-binary==2.9.5
requests==2.25.1
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
zipp==3.4.1